+++ CMakeLists.txt
@@ -257,18 +257,6 @@ execute_process(COMMAND ${CMAKE_COMMAND} -E copy_if_different
   ${CMAKE_CURRENT_BINARY_DIR}/translations.qrc)
 
 set(${PROJECT_NAME}_SOURCES
-  src/audio/audio.cpp
-  src/audio/audio.h
-  src/audio/backend/alsink.cpp
-  src/audio/backend/alsink.h
-  src/audio/backend/alsource.cpp
-  src/audio/backend/alsource.h
-  src/audio/backend/openal.cpp
-  src/audio/backend/openal.h
-  src/audio/iaudiosettings.h
-  src/audio/iaudiocontrol.h
-  src/audio/iaudiosink.h
-  src/audio/iaudiosource.h
   src/chatlog/chatlinecontent.cpp
   src/chatlog/chatlinecontent.h
   src/chatlog/chatlinecontentproxy.cpp
@@ -368,7 +356,6 @@ set(${PROJECT_NAME}_SOURCES
   src/model/group.h
   src/model/status.cpp
   src/model/status.h
-  src/model/interface.h
   src/model/profile/iprofileinfo.h
   src/model/profile/profileinfo.cpp
   src/model/profile/profileinfo.h
@@ -659,6 +646,9 @@ MESSAGE( STATUS "CMAKE_C_FLAGS: " ${CMAKE_C_FLAGS} )
 # the compiler flags for compiling C++ sources
 MESSAGE( STATUS "CMAKE_CXX_FLAGS: " ${CMAKE_CXX_FLAGS} )
 
+add_subdirectory(util)
+add_subdirectory(audio)
+
 add_library(${PROJECT_NAME}_static
   STATIC
   ${${PROJECT_NAME}_FORMS}
@@ -668,6 +658,9 @@ add_library(${PROJECT_NAME}_static
 target_link_libraries(${PROJECT_NAME}_static
   ${CMAKE_REQUIRED_LIBRARIES}
   ${ALL_LIBRARIES})
+
+target_link_libraries(${PROJECT_NAME}_static util_library)
+target_link_libraries(${PROJECT_NAME}_static audio_library)
 
 add_executable(${PROJECT_NAME}
   WIN32
