+++ audio/include/audio/iaudiosink.h
@@ -23,7 +23,7 @@
 
 #include <QObject>
 
-#include "src/model/interface.h"
+#include "util/interface.h"
 
 /**
  * @brief The IAudioSink class represents an interface to devices that can play audio.
