+++ audio/src/backend/alsink.h
@@ -22,8 +22,8 @@
 #include <QMutex>
 #include <QObject>
 
-#include "src/model/interface.h"
-#include "src/audio/iaudiosink.h"
+#include "util/interface.h"
+#include "audio/iaudiosink.h"
 
 class OpenAL;
 class QMutex;
