+++ src/model/about/aboutfriend.h
@@ -20,7 +20,7 @@
 #pragma once
 
 #include "iaboutfriend.h"
-#include "src/model/interface.h"
+#include "util/interface.h"
 #include "src/persistence/ifriendsettings.h"
 
 #include <QObject>
