+++ src/model/about/iaboutfriend.h
@@ -19,7 +19,7 @@
 
 #pragma once
 
-#include "src/model/interface.h"
+#include "util/interface.h"
 #include "src/persistence/ifriendsettings.h"
 
 #include <QObject>
