+++ src/model/profile/profileinfo.h
@@ -20,7 +20,7 @@
 #pragma once
 
 #include <QObject>
-#include "src/model/interface.h"
+#include "util/interface.h"
 #include "src/core/toxpk.h"
 #include "iprofileinfo.h"
