+++ util/CMakeLists.txt
@@ -1,0 +1,25 @@
+#   Copyright © 2020 by The qTox Project Contributors
+#
+#   This file is part of qTox, a Qt-based graphical interface for Tox.
+#   qTox is libre software: you can redistribute it and/or modify
+#   it under the terms of the GNU General Public License as published by
+#   the Free Software Foundation, either version 3 of the License, or
+#   (at your option) any later version.
+#
+#   qTox is distributed in the hope that it will be useful,
+#   but WITHOUT ANY WARRANTY; without even the implied warranty of
+#   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#   GNU General Public License for more details.
+#
+#   You should have received a copy of the GNU General Public License
+#   along with qTox.  If not, see <http://www.gnu.org/licenses/>
+
+set(HEADER_LIST
+    "include/util/interface.h"
+    "include/util/strongtype.h")
+
+add_library(util_library STATIC ${HEADER_LIST})
+
+# We need this directory, and users of our library will need it too
+target_include_directories(util_library PUBLIC include/)
+
